backend/requirements.txt
python-jose[cryptography]==3.3.0

bcrypt==4.0.1
scikit-learn>=1.5.0
numpy>=2.0.0
spacy>=3.7.0
# axios
python-dotenv==1.0.0
redis==5.0.1
bleach==6.1.0
prometheus-client==0.19.0

psutil==5.9.6
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy>=2.0.25
mysql-connector-python==8.2.0
alembic==1.12.1
sentence-transformers==2.2.2
faiss-cpu
openai==1.3.0
